move mixins accessor interfaces under mixins\interfaces package
